+++ src/spanner/types.ts
@@ -1,3 +1,8 @@
+import type {
+  SpannerReadOnlyTransaction,
+  SpannerReadWriteTransaction,
+} from './entity-manager.js';
+
 /**
  * A partial Spanner entity instance, where nested objects can also be partial.
  */
@@ -6,3 +11,23 @@ export type RecursivePartialEntity<T> = T extends Date
   : T extends object
     ? Partial<T> | { [P in keyof T]?: RecursivePartialEntity<T[P]> }
     : T;
+
+/**
+ * Option for a function that accepts a Spanner read-only transaction.
+ */
+export type SpannerReadOnlyTransactionOption = {
+  /**
+   * The transaction to use.
+   */
+  readonly transaction?: SpannerReadOnlyTransaction;
+};
+
+/**
+ * Option for a function that accepts a Spanner read-write transaction.
+ */
+export type SpanerReadWriteTransactionOption = {
+  /**
+   * The transaction to use.
+   */
+  readonly transaction?: SpannerReadWriteTransaction;
+};